Grouping views events in the ViewEventArgs.cs file.
