+++ Terminal.Gui/Core/ViewEventArgs.cs
@@ -1,17 +1,13 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Terminal.Gui {
 	/// <summary>
 	/// Args for events that relate to specific <see cref="View"/>
 	/// </summary>
-	public class ViewEventArgs :EventArgs{
+	public class ViewEventArgs : EventArgs {
 
 		/// <summary>
-		/// Creates a new instance of the <see cref="ViewEventArgs"/> class.
+		/// Creates a new instance of the <see cref="Terminal.Gui.View"/> class.
 		/// </summary>
 		/// <param name="view"></param>
 		public ViewEventArgs (View view)
@@ -30,4 +26,55 @@ namespace Terminal.Gui {
 		/// </remarks>
 		public View View { get; }
 	}
+
+	/// <summary>
+	/// Event arguments for the <see cref="View.LayoutComplete"/> event.
+	/// </summary>
+	public class LayoutEventArgs : EventArgs {
+		/// <summary>
+		/// The view-relative bounds of the <see cref="View"/> before it was laid out.
+		/// </summary>
+		public Rect OldBounds { get; set; }
+	}
+
+	/// <summary>
+	/// Event args for draw events
+	/// </summary>
+	public class DrawEventArgs : EventArgs {
+
+		/// <summary>
+		/// Creates a new instance of the <see cref="DrawEventArgs"/> class.
+		/// </summary>
+		/// <param name="rect">Gets the view-relative rectangle describing the currently visible viewport into the <see cref="View"/>.</param>
+		public DrawEventArgs (Rect rect)
+		{
+			Rect = rect;
+		}
+
+		/// <summary>
+		/// Gets the view-relative rectangle describing the currently visible viewport into the <see cref="View"/>.
+		/// </summary>
+		public Rect Rect { get; }
+	}
+
+	/// <summary>
+	/// Defines the event arguments for <see cref="View.SetFocus(View)"/>
+	/// </summary>
+	public class FocusEventArgs : EventArgs {
+		/// <summary>
+		/// Constructs.
+		/// </summary>
+		/// <param name="view">The view that gets or loses focus.</param>
+		public FocusEventArgs (View view) { View = view; }
+		/// <summary>
+		/// Indicates if the current focus event has already been processed and the driver should stop notifying any other event subscriber.
+		/// Its important to set this value to true specially when updating any View's layout from inside the subscriber method.
+		/// </summary>
+		public bool Handled { get; set; }
+		/// <summary>
+		/// Indicates the current view that gets or loses focus.
+		/// </summary>
+		public View View { get; set; }
+	}
+
 }
